+++ src/features/studynotes/services/query.ts
@@ -1,7 +1,7 @@
 import { Role } from '@/features/auth/type';
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
-import type { StudyNoteGroupPageable } from '../../studyrooms/components/studynotes/type';
+import type { StudyNoteGroupPageable } from '../type';
 import { deleteStudyNoteToGroup, updateStudyNoteGroup } from './api';
 import { updateStudyNote } from './api';
 import {
